improve SignColumn, add nice ShowMarks support
The SignColumn which appears to the left of the line number
column stood out in a very bad way because it was white.
I changed it to have the same appearance as the line number
column by using color linking. I also linked the ShowMarks
to DiffAdd and DiffChanged because those were nice already.